#399: lift similarity helpers into shared utils.pattern_similarity
canonicalize_for_dedupe + similarity + the dedupe bands move to
utils/pattern_similarity.py so the community validator and the runtime
merge/clustering paths use one implementation. Validator re-exports them.
